Split out Python 3.X test.
